+++ Demo/Demo/Models/PaymentTokenRequest.swift
@@ -1,5 +1,4 @@
 import Foundation
-import CorePayments
 
 struct PaymentTokenRequest {
     
@@ -8,13 +7,13 @@ struct PaymentTokenRequest {
     var path: String {
         "/payment_tokens/"
     }
-    
-    var method: CorePayments.HTTPMethod {
-        .post
+
+    var method: String {
+        "POST"
     }
     
-    var headers: [CorePayments.HTTPHeader: String] {
-        [.contentType: "application/json"]
+    var headers: [String: String] {
+        ["Content-Type": "application/json"]
     }
     
     var body: Data? {
@@ -26,6 +25,7 @@ struct PaymentTokenRequest {
                 ]
             ]
         ]
+
         return try? JSONSerialization.data(withJSONObject: requestBody)
     }
 }