+++ Demo/Demo/Networking/DemoMerchantAPI.swift
@@ -26,7 +26,9 @@ final class DemoMerchantAPI {
             // in SetUpTokenRequest to conditionally add header
             let request = SetUpTokenRequest(customerID: customerID)
             let urlRequest = try createSetupTokenUrlRequest(
-                setupTokenRequest: request, environment: DemoSettings.environment, selectedMerchantIntegration: selectedMerchantIntegration
+                setupTokenRequest: request,
+                environment: DemoSettings.environment,
+                selectedMerchantIntegration: selectedMerchantIntegration
             )
             
             let data = try await data(for: urlRequest)
@@ -36,16 +38,16 @@ final class DemoMerchantAPI {
             throw error
         }
     }
-    
-	func getPaymentToken(setupToken: String, selectedMerchantIntegration: MerchantIntegration) async throws -> PaymentTokenResponse {
+
+    func getPaymentToken(setupToken: String, selectedMerchantIntegration: MerchantIntegration) async throws -> PaymentTokenResponse {
         do {
             let request = PaymentTokenRequest(setupToken: setupToken)
             let urlRequest = try createPaymentTokenUrlRequest(
                 paymentTokenRequest: request,
                 environment: DemoSettings.environment,
                 selectedMerchantIntegration: selectedMerchantIntegration
-	)
-		let data = try await data(for: urlRequest)
+			)
+			let data = try await data(for: urlRequest)
             return try parse(from: data)
         } catch {
             print("error with the create payment token request: \(error.localizedDescription)")
@@ -228,18 +230,20 @@ final class DemoMerchantAPI {
         selectedMerchantIntegration: MerchantIntegration
     ) throws -> URLRequest {
         var completeUrl = environment.baseURL
-       
         completeUrl += selectedMerchantIntegration.path
         completeUrl.append(contentsOf: setupTokenRequest.path)
+
         guard let url = URL(string: completeUrl) else {
             throw URLResponseError.invalidURL
         }
+
         var request = URLRequest(url: url)
-        request.httpMethod = setupTokenRequest.method.rawValue
+        request.httpMethod = setupTokenRequest.method
         request.httpBody = setupTokenRequest.body
         setupTokenRequest.headers.forEach { key, value in
-            request.addValue(value, forHTTPHeaderField: key.rawValue)
+            request.addValue(value, forHTTPHeaderField: key)
         }
+
         return request
     }
     
@@ -249,18 +253,20 @@ final class DemoMerchantAPI {
         selectedMerchantIntegration: MerchantIntegration
     ) throws -> URLRequest {
         var completeUrl = environment.baseURL
-       
         completeUrl += selectedMerchantIntegration.path
         completeUrl.append(contentsOf: paymentTokenRequest.path)
+
         guard let url = URL(string: completeUrl) else {
             throw URLResponseError.invalidURL
         }
+        
         var request = URLRequest(url: url)
-        request.httpMethod = paymentTokenRequest.method.rawValue
+        request.httpMethod = paymentTokenRequest.method
         request.httpBody = paymentTokenRequest.body
         paymentTokenRequest.headers.forEach { key, value in
-            request.addValue(value, forHTTPHeaderField: key.rawValue)
+            request.addValue(value, forHTTPHeaderField: key)
         }
+
         return request
     }
 }