+++ Sources/CorePayments/WebAuthenticationSession.swift
@@ -15,7 +15,6 @@ public class WebAuthenticationSession: NSObject {
             completionHandler: sessionDidComplete
         )
 
-        authenticationSession.prefersEphemeralWebBrowserSession = true
         authenticationSession.presentationContextProvider = context
 
         DispatchQueue.main.async {