+++ UnitTests/PaymentsCoreTests/GraphQLClient_Tests.swift
@@ -14,7 +14,8 @@ class GraphQLClient_Tests: XCTestCase {
     var mockURLSession: MockURLSession!
     var graphQLClient: GraphQLClient!
     var graphQLQuery: (any GraphQLQuery)?
-    // TODO: add unit tests in appropriate class after refactor is done
+
+    // TODO: add unit tests in appropriate class after networking & graphQL layer refactor is done
 
     // MARK: - Test lifecycle